Cleanup: create runtime1 package and mixins for Term classes and Gram.
The purpose here is separation of concerns.
Things specific to daffodil-runtime1 should be segregated in the schema
compiler to a package with that specific runtime/backend in mind.

This makes it clear that the dsom and grammar packages are supposed to contain
backend-independent code.

However, this is a very partial first step in this direction. The
grammar primitives are still very runtime1-specific with their parser and
unparser methods. Also the dpath package needs separating as well.

Nevertheless, this change is worth it to help better categorize the
various schema compiler optimizers and various computed LV "attributes" into
those that are runtime1 specific, and those that are fully general.

DAFFODIL-2233
